+++ fabric/src/main/java/nl/curryducker/seamless/fabric/mixin/RiceUpperCropBlockMixin.java
@@ -1,13 +1,11 @@
 package nl.curryducker.seamless.fabric.mixin;
 
-import net.mehvahdjukaar.moonlight.api.block.IBeeGrowable;
-import net.mehvahdjukaar.supplementaries.common.block.blocks.FlaxBlock;
+import com.nhoryzon.mc.farmersdelight.block.RiceUpperCropBlock;
 import net.minecraft.core.BlockPos;
 import net.minecraft.world.level.BlockGetter;
 import net.minecraft.world.level.block.CropBlock;
 import net.minecraft.world.level.block.state.BlockState;
-import net.minecraft.world.level.block.state.properties.DoubleBlockHalf;
-import net.minecraft.world.level.block.state.properties.EnumProperty;
+import net.minecraft.world.level.block.state.properties.IntegerProperty;
 import net.minecraft.world.phys.shapes.CollisionContext;
 import net.minecraft.world.phys.shapes.VoxelShape;
 import nl.curryducker.seamless.SeamlessShapes;
@@ -18,18 +16,17 @@ import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
-@Mixin(FlaxBlock.class)
-public abstract class FlaxBlockMixin extends CropBlock implements IBeeGrowable {
-    @Shadow @Final public static EnumProperty<DoubleBlockHalf> HALF;
+@Mixin(RiceUpperCropBlock.class)
+public abstract class RiceUpperCropBlockMixin extends CropBlock {
 
-    public FlaxBlockMixin(Properties properties) {
+    @Shadow @Final public static IntegerProperty RICE_AGE;
+
+    public RiceUpperCropBlockMixin(Properties properties) {
         super(properties);
     }
 
     @Inject(method = "getShape", at = @At("HEAD"), cancellable = true)
     public void getShape(BlockState state, BlockGetter worldIn, BlockPos pos, CollisionContext context, CallbackInfoReturnable<VoxelShape> cir) {
-        if (state.getValue(AGE)>=4) {
-            cir.setReturnValue(SeamlessShapes.flax(state.getValue(HALF) == DoubleBlockHalf.LOWER, state.getValue(AGE)));
-        }
+        cir.setReturnValue(SeamlessShapes.rice(3, state.getValue(RICE_AGE), false));
     }
 }