+++ forge/src/main/java/nl/curryducker/seamless/forge/mixin/DoubleCattailBlockMixin.java
@@ -1,36 +1,35 @@
 package nl.curryducker.seamless.forge.mixin;
 
-import net.mehvahdjukaar.sleep_tight.common.blocks.IModBed;
-import net.mehvahdjukaar.sleep_tight.common.blocks.NightBagBlock;
+import com.teamabnormals.environmental.common.block.DoubleCattailBlock;
 import net.minecraft.core.BlockPos;
-import net.minecraft.world.item.DyeColor;
 import net.minecraft.world.level.BlockGetter;
-import net.minecraft.world.level.block.BedBlock;
 import net.minecraft.world.level.block.Block;
+import net.minecraft.world.level.block.BonemealableBlock;
+import net.minecraft.world.level.block.SimpleWaterloggedBlock;
 import net.minecraft.world.level.block.state.BlockState;
-import net.minecraft.world.level.block.state.properties.BedPart;
+import net.minecraft.world.level.block.state.properties.DoubleBlockHalf;
+import net.minecraft.world.level.block.state.properties.EnumProperty;
 import net.minecraft.world.phys.shapes.CollisionContext;
 import net.minecraft.world.phys.shapes.VoxelShape;
 import nl.curryducker.seamless.SeamlessShapes;
+import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
-@Mixin(NightBagBlock.class)
-public abstract class NightBagBlockMixin extends BedBlock implements IModBed {
+@Mixin(DoubleCattailBlock.class)
+public abstract class DoubleCattailBlockMixin extends Block implements BonemealableBlock, SimpleWaterloggedBlock {
 
-    public NightBagBlockMixin(DyeColor dyeColor, Properties properties) {
-        super(dyeColor, properties);
+    public DoubleCattailBlockMixin(Properties properties) {
+        super(properties);
     }
+
+    @Shadow @Final public static EnumProperty<DoubleBlockHalf> HALF;
 
     @Inject(method = "getShape", at = @At("HEAD"), cancellable = true)
     public void getShape(BlockState state, BlockGetter worldIn, BlockPos pos, CollisionContext context, CallbackInfoReturnable<VoxelShape> cir) {
-        cir.setReturnValue(SeamlessShapes.mat(state.getValue(FACING), state.getValue(PART) == BedPart.FOOT));
-    }
-
-    @Override
-    public VoxelShape getCollisionShape(BlockState blockState, BlockGetter blockGetter, BlockPos blockPos, CollisionContext collisionContext) {
-        return Block.box(0.0, 0.0, 0.0, 16.0, 2.0, 16.0);
+        cir.setReturnValue(SeamlessShapes.cattail(state.getValue(HALF) == DoubleBlockHalf.LOWER));
     }
 }